Separated form schema for login and sign up
